Raspb5/requirements.txt
selenium
webdriver-manager
pyserial
